# src/openapi.yaml
openapi: 3.0.0
info:
    title: Space bots
    version: 1.0.0
servers:
    - url: "{host}/v1/"
      description: Main server
      variables:
          host:
              default: https://space-bots.longwelwind.net
security:
    - bearerAuth: []
paths:
    /users/me:
        get:
            summary: Returns information about yourself
            tags:
                - Users
            responses:
                "200":
                    description: Info about yourself
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    id:
                                        type: string
                                    name:
                                        type: string
                                    credits:
                                        type: integer
                                    registeredAt:
                                        type: string
                                        format: date-time
                                    registered:
                                        type: boolean
                                        description: Whether this user has already registered their username if they logged in via a third-party provider
                "401":
                    $ref: "#/components/responses/401_unauthorized"
                "429":
                    $ref: "#/components/responses/429_too_many_requests"
    /users/register:
        post:
            summary: Register yourself
            tags:
                - Users
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                name:
                                    type: string
            responses:
                "200":
                    description: TODO
                "400":
                    description: Error
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: string
                                        enum: ["name_taken"]
                "401":
                    $ref: "#/components/responses/401_unauthorized"
                "429":
                    $ref: "#/components/responses/429_too_many_requests"
    /systems/{systemId}:
        get:
            summary: Return info about a system
            tags:
                - Systems
            parameters:
                - $ref: "#/components/parameters/systemId"
            responses:
                "200":
                    description: Info about the system
                    content:
                        application/json:
                            schema:
                                type: object
                                required:
                                    - id
                                    - name
                                    - x
                                    - y
                                    - neighboringSystems
                                properties:
                                    id:
                                        type: string
                                    name:
                                        type: string
                                    asteroid:
                                        type: object
                                        required:
                                            - miningResourceId
                                        properties:
                                            miningResourceId:
                                                type: string
                                    station:
                                        type: object
                                        required:
                                            - buyShips
                                            - directSell
                                        properties:
                                            buyShips:
                                                type: boolean
                                                description: Whether you can buy ships at this station
                                            directSell:
                                                type: boolean
                                                description: Whether you can sell resources directly at this station
                                    x:
                                        type: number
                                    y:
                                        type: number
                                    neighboringSystems:
                                        type: array
                                        description: List of IDs of the systems
                                        items:
                                            type: object
                                            required:
                                                - systemId
                                            properties:
                                                systemId:
                                                    type: string
                "401":
                    $ref: "#/components/responses/401_unauthorized"
                "404":
                    description: System not found
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: string
                                        enum: ["not_found"]
                "429":
                    $ref: "#/components/responses/429_too_many_requests"
...
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
    responses:
        "401_unauthorized":
            description: No credentials or invalid credentials
            content:
                application/json:
                    schema:
                        type: object
                        properties:
                            error:
                                type: string
                                enum: ["unauthorized"]
        "429_too_many_requests":
            description: Too many requests were sent
            content:
                application/json:
                    schema:
                        type: object
                        properties:
                            error:
                                type: string
                                enum: ["rate_limited"]
    parameters:
...
        systemId:
            in: path
            name: systemId
            schema:
                type: string
            required: true
            description: ID of the system
...